# src/main/resources/spec/api.yml
openapi: 3.0.1
info:
  title: Swagger Technight Api First
  description: |-
    Api qui permets de récupérer des données depuis l'API TheMealDB
  contact:
    email: fguibon@oxyl.fr
  version: 1.0.0
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: http://localhost:8080
tags:
  - name: Meals
    description: Search meals
paths:
  /meals:
    get:
      tags:
        - Meals
      summary: Search for meals by ingredient, category or area
      operationId: getMeals
      parameters:
        - name: ingredient
          in: query
          description: Main ingredient name
          required: false
          schema:
            type: string
        - name: category
          in: query
          description: Category
          required: false
          schema:
            type: string
        - name: area
          in: query
          description: Area
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MealList'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /meals/{id}:
    get:
      tags:
        - Meals
      summary: Get meal by id
      operationId: getMealById
      parameters:
        - name: id
          in: path
          description: Id of the meal
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MealDetailed'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          example: "Invalid parameters"
    Meal:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "Katsu Chicken curry"
        thumbnail:
          type: string
          example: "https://www.themealdb.com/images/media/meals/vwrpps1503068729.jpg"
    MealList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Meal'
    MealDetailed:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 52820
        name:
          type: string
          example: "Katsu Chicken curry"
        category:
          type: string
          example: "Chicken"
        area:
          type: string
          example: "Japanese"
        thumbnail:
          type: string
          example: "https://www.themealdb.com/images/media/meals/vwrpps1503068729.jpg"
        instructions:
          type: string
          example: "Prep:15min  ›  Cook:30min  ›  Ready in:45min \r\n\r\nFor the curry sauce: Heat oil in medium non-stick saucepan, add onion and garlic and cook until softened. Stir in carrots and cook over low heat for 10 to 12 minutes.\r\nAdd flour and curry powder; cook for 1 minute. Gradually stir in stock until combined; add honey, soy sauce and bay leaf. Slowly bring to the boil.\r\nTurn down heat and simmer for 20 minutes or until sauce thickens but is still of pouring consistency. Stir in garam masala. Pour the curry sauce through a sieve; return to saucepan and keep on low heat until ready to serve.\r\nFor the chicken: Season both sides of chicken breasts with salt and pepper. Place flour, egg and breadcrumbs in separate bowls and arrange in a row. Coat the chicken breasts in flour, then dip them into the egg, then coat in breadcrumbs, making sure you cover both sides.\r\nHeat oil in large frying pan over medium-high heat. Place chicken into hot oil and cook until golden brown, about 3 or 4 minutes each side. Once cooked, place on kitchen paper to absorb excess oil.\r\nPour curry sauce over chicken, serve with white rice and enjoy!"
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
        source:
          type: string
          example: "http://allrecipes.co.uk/recipe/29578/chicken-katsu-curry.aspx"
    Ingredient:
      type: object
      properties:
        name:
          type: string
          example: "chicken breast"
        measure:
          type: string
          example: "4 pounded to 1cm thickness"